+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - (Feature) Recreation member in the high plan
 - (Feature) Add 'crd install' subcommand
 - (Bugfix) Fix `internal` metrics mode
+- (Bugfix) Create agency dump if auth is disabled
 
 ## [1.2.14](https://github.com/arangodb/kube-arangodb/tree/1.2.14) (2022-07-14)
 - (Feature) Add ArangoSync TLS based rotation